refactor: move outer testing modules to the inner
